Synchronize codes for OxygenOS 9.0
Change list:
  - System
    - Updated system to Android 9 Pie
    - Brand new UI for Android Pie
    - Updated Android security patch to 2019.4

+++ drivers/nfc/pn544_op/pn544.h
@@ -68,7 +68,7 @@
 /*
   NFC Init will call the ioctl to register the PID with the i2c driver
 */
-#define P544_SET_NFC_SERVICE_PID _IOW(PN544_MAGIC, 0x05, long)
+#define P544_SET_NFC_SERVICE_PID _IOW(PN544_MAGIC, 0x05, unsigned int)
 
 typedef enum p61_access_state{
     P61_STATE_INVALID = 0x0000,
